Restructured 'attachments' by applying sealed class
